fix: move router setup to the story instance
