+++ stonecrop/src/index.js
@@ -2,12 +2,12 @@ import router from './router.js'
 import Registry from './registry.js'
 // import { useStonecrop } from './composable.js'
 
-// import { pinia } from "./store.js"
+// TODO: import { pinia } from "./store.js"
 
 const Stonecrop = {
 	install: (app, options) => {
 		app.use(router)
-		app.provide('$registry', new Registry(router, options.schemaLoader))
+		app.provide('$registry', new Registry(options.router || router, options.schemaLoader))
 	},
 }
